+++ aiapy/calibrate/util.py
@@ -13,12 +13,12 @@ import numpy as np
 from astropy.table import QTable
 from astropy.time import Time
 from erfa.core import ErfaWarning
+from sunpy import log
 from sunpy.data import manager
 from sunpy.net import attrs, jsoc
 
 from aiapy import _SSW_MIRRORS
 from aiapy.util.decorators import validate_channel
-from aiapy.util.exceptions import AiapyUserWarning
 
 __all__ = ["get_correction_table", "get_pointing_table", "get_error_table"]
 
@@ -48,11 +48,11 @@ def get_correction_table(*, correction_table=None):
 
     This function returns a table of parameters for estimating the
     time-dependent degradation of the instrument. By default, this table
-    is queried from `aia.response` series in
-    `JSOC <http://jsoc.stanford.edu/>`_. The correction table can also be read
-    from a file by passing a filepath to `correction_table`. These files are
+    is queried from ``aia.response`` series in
+    `JSOC <http://jsoc.stanford.edu/>`__. The correction table can also be read
+    from a file by passing a filepath to ``correction_table``. These files are
     typically included in the SDO tree of an SSW installation in
-    `$SSW/sdo/aia/response/` with filenames like `aia_V*_response_table.txt`.
+    ``$SSW/sdo/aia/response/`` with filenames like ``aia_V*_response_table.txt``.
 
     Parameters
     ----------
@@ -116,7 +116,7 @@ def get_correction_table(*, correction_table=None):
 def _select_epoch_from_correction_table(channel: u.angstrom, obstime, table, *, version=None):
     """
     Return correction table with only the first epoch and the epoch in which
-    `obstime` falls and for only one given calibration version.
+    ``obstime`` falls and for only one given calibration version.
 
     Parameters
     ----------
@@ -147,10 +147,8 @@ def _select_epoch_from_correction_table(channel: u.angstrom, obstime, table, *, 
     # use the most up-to-date one.
     i_epoch = np.where(obstime_in_epoch)[0]
     if i_epoch.shape[0] > 1:
-        warnings.warn(
+        log.debug(
             f"Multiple valid epochs for {obstime}. Using the most recent one",
-            AiapyUserWarning,
-            stacklevel=3,
         )
     # Create new table with only first and obstime epochs
     return QTable(table[[0, i_epoch[-1]]])
@@ -162,7 +160,7 @@ def get_pointing_table(start, end):
 
     This function queries `JSOC <http://jsoc.stanford.edu/>`__ for
     the 3-hourly master pointing table (MPT) in the interval defined by
-    `start` and `end`.
+    ``start`` and ``end``.
     The 3-hourly MPT entries are computed from limb fits of images with
     ``T_OBS`` between ``T_START`` and ``T_STOP``.